apps/ai-engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pgvector==0.2.4
sqlalchemy==2.0.25
redis==5.0.1
openai==1.7.2
langchain==0.1.0
langchain-openai==0.0.2
numpy==1.26.3
scikit-learn==1.4.0
python-multipart==0.0.6
